+++ changelog.md
@@ -1,17 +1,20 @@
-# gooseRSS
+# GooseRSS
 
 Versioning is loose and lax, in fact there are no actual versions.  
 But here is the list of changes made over time, sorted by 'release' date.
 
-## April 2, 2026
+## April 6, 2026
 - Fix: Consistent use trailing / for MAIN_URL
 - New: When adding a feed a welcome item is added as a placeholder
 - New: Successful deletion of old cache files logged in success.log 
+- New: MAIN_PATH must be defined for included files to work
 - Change: Feeds now keep up to 50 items in cache
+- Change: moved functions.php to /functions/functions.php
+- Change: index.php files replaced with index.html
 - Change: Old (and orphaned) cache files cleared once a month around 11AM GMT
+- Change: check_config now also checks for index.html in cache
 - Rollback: Feeds silently fail again if any error occurs while fetching new data
 - Rollback: Removed ERROR_FEED from default-config.php
-- Remove: Ability to skip "Live" and "Premiere" videos since it's nearly undetectable from XML alone
 
 ## April 1, 2026
 - Update: Refined cache cleanup routine